added type checking of params in function calls

+++ test/test30.c
@@ -63,7 +63,7 @@ int main()
             prints("Error at offset ");
             printi(index);
             prints(" ");
-            printi(datap[index]);
+            printi((int)datap[index]);
             prints("\n");
         }
         else
